# Update Lib: xor -> xorsimd, slight perf improvement