➜  kcp-go git:(refine/xor2) ✗ benchcmp old.txt new.txt
benchmark                        old ns/op     new ns/op     delta
BenchmarkAES128-8                2777          2581          -7.06%
BenchmarkAES192-8                3067          2874          -6.29%
BenchmarkAES256-8                3300          3201          -3.00%
BenchmarkCsprngNonceAES128-8     15.5          15.3          -1.29%

benchmark                        old MB/s     new MB/s     speedup
BenchmarkAES128-8                1080.19      1162.22      1.08x
BenchmarkAES192-8                978.28       1043.86      1.07x
BenchmarkAES256-8                909.06       937.27       1.03x
BenchmarkCsprngNonceAES128-8     1033.62      1046.87      1.01x

benchmark             old allocs     new allocs     delta
BenchmarkAES128-8     0              0              +0.00%
BenchmarkAES192-8     0              0              +0.00%
BenchmarkAES256-8     0              0              +0.00%

benchmark             old bytes     new bytes     delta
BenchmarkAES128-8     0             0             +0.00%
BenchmarkAES192-8     0             0             +0.00%
BenchmarkAES256-8     0             0             +0.00%